requirements.txt
Flask==2.3.3
slack-sdk==3.21.3
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1      
google-auth-oauthlib==1.1.0      
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
cryptography==41.0.7 